Moved CpgLoader and overlay loading related classes to codepropertygraph.
