+++ query-primitives/build.sbt
@@ -4,7 +4,6 @@ dependsOn(Projects.codepropertygraph)
 
 libraryDependencies ++= Seq(
   "org.json4s" %% "json4s-native" % "3.5.2",
-  "org.scala-lang.modules" %% "scala-java8-compat" % "0.8.0",
   "com.massisframework" % "j-text-utils" % "0.3.4",
   "org.scalatest" %% "scalatest" % "3.0.3" % Test
 )
